requirements.txt
boto3>=1.26.0
great-expectations>=0.15.0
pyspark>=3.3.0
pyyaml>=6.0
pandas>=1.5.0
numpy>=1.23.0
pytest>=7.0.0
black>=22.0.0
flake8>=6.0.0
mypy>=1.0.0 